UPDATE:优化

+++ app/Plugins/User/src/migrations/2022_08_13_190547_update_users_remove_fans.php
@@ -12,7 +12,9 @@ class UpdateUsersRemoveFans extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('fans');
+            if (Schema::hasColumn('users', 'fans')) {
+                $table->dropColumn('fans');
+            }
         });
     }
